+++ Process.h
@@ -380,7 +380,7 @@ void Process_fillStarttimeBuffer(Process* this);
 
 void Process_printLeftAlignedField(RichString* str, int attr, const char* content, unsigned int width);
 
-void Process_printPercentage(float val, char* buffer, int n, uint8_t width, int* attr);
+void Process_printPercentage(float val, char* buffer, size_t n, uint8_t width, int* attr);
 
 void Process_display(const Object* cast, RichString* out);
